Added proper support for environment variables and CWD.
This is a bit tricky due to the different targets and languages we
support. On host and with C++, we have getenv etc. from the host libc,
but cannot use it to prevent that we interfere with the host. On kachel
and with C++, we have getenv from musl. With Rust, we don't have std
and therefore no support for env vars at all.

To support all target and languages properly, I've decided to provide my
own implementation in C++ and Rust. In C++ we have to be careful to be
compatible, so that we can use the M³ API and musl API interchangeably.

Based on that, we provide VFS::cwd/VFS::set_cwd as well and respect the
current working directory for all VFS methods. The shell also fully
supports relative paths now and has got builtin commands to support cd
(and also echo).

+++ src/libs/m3/arch/host/Activity.cc
@@ -21,13 +21,14 @@
 #include <base/Panic.h>
 #include <base/Init.h>
 
+#include <m3/EnvVars.h>
 #include <m3/session/ResMng.h>
 #include <m3/stream/FStream.h>
+#include <m3/Syscalls.h>
+#include <m3/tiles/Activity.h>
 #include <m3/vfs/FileTable.h>
 #include <m3/vfs/MountTable.h>
 #include <m3/vfs/VFS.h>
-#include <m3/Syscalls.h>
-#include <m3/tiles/Activity.h>
 
 #include <sys/fcntl.h>
 #include <sys/stat.h>
@@ -111,7 +112,10 @@ static void *read_from(const char *suffix, void *dst, size_t &size) {
             dst = malloc(size);
         }
 
-        read(fd, dst, size);
+        ssize_t res = read(fd, dst, size);
+        if(res < 0)
+            return nullptr;
+        size = static_cast<size_t>(res);
         unlink(path);
         close(fd);
         return dst;
@@ -180,6 +184,21 @@ void OwnActivity::init_fs() {
 
     len = sizeof(_data);
     read_from("data", _data, len);
+
+    if(read_from("vars", buf.get(), len)) {
+        Unmarshaller um(reinterpret_cast<unsigned char*>(&*buf.get()), len);
+        while(um.remaining() > 0) {
+            String s;
+            um >> s;
+            size_t pos = static_cast<size_t>(strchr(s.c_str(), '=') - s.c_str());
+            char *key = static_cast<char*>(malloc(pos + 1));
+            assert(key != nullptr);
+            strncpy(key, s.c_str(), pos);
+            key[pos] = '\0';
+            EnvVars::set(key, s.c_str() + pos + 1);
+            free(key);
+        }
+    }
 
     // TCU is ready now; notify parent
     int pipefd;
@@ -283,6 +302,11 @@ void ChildActivity::do_exec(int argc, const char **argv, uintptr_t func_addr) {
 
         len = Activity::own().files()->serialize(*this, buf.get(), STATE_BUF_SIZE);
         write_file(pid, "fds", buf.get(), len);
+
+        Marshaller m(buf.get(), STATE_BUF_SIZE);
+        for(size_t i = 0; i < EnvVars::count(); ++i)
+            m << EnvVars::vars()[i];
+        write_file(pid, "vars", buf.get(), m.total());
 
         write_file(pid, "data", _data, sizeof(_data));
